Refine middleware and add a basic RequestLogger middleware
A middleware will be a before, around, or after depending upon how it
continues the execution chain in relation to its functionality.

The RequestLogger is an example of an *around* middleware, since it
starts (by capturing the time) before executing the handler it was given
and then by logging afterwards.

A `Status` was also added to the context so that middleware have access
to the outcome of any given response.