Update topic/web/{basics,testing} to better explain basic http types
The flow of the examples is to first introduce http.Handler and
http.ListenAndServe. Next we create multiple handlers that can be routed
using a *http.ServeMux and then the default mux. Creating a new type for
each route is tedious so we then introduce the idea of having one type
with multiple methods and wiring them up using HandleFunc on a mux.
Finally we show how any func with the right signature can become a
handler thanks to the http.HandlerFunc adapter type.

Oh and I also restructured the concurrency example to be less clever.