topics/data: Move most gonum imports to new path
The github.com/gonum imports have been deprecated in favor of their
gonum.org equivalents. I have updated most of the references but a few
of the mat64 examples still require the old version because of a
transitive dependency from the gota package. A comment has been added to
these where necessary.