Switch github.com/theplant/blackfriday to pull from master branch
When `dep` was ran on this repo last (d1506e457d), `dep` decided to use `v1.1.0`
of `github.com/theplant/blackfriday`. Normally this selection would have been
correct, but there is a bug in v1.1.0 that's been fixed in master, with no git
tag including the fix.

Specifically, one of the files tried to import `utf8` instead of `unicode/utf8`.
The fix is to change `Gopkg.toml` to use the `master` branch of
`github.com/theplant/blackfriday` and to then run `dep ensure`.

Fixes #244

Signed-off-by: Tim Heckman <t@heckman.io>