pointers: Show how Go will conveniently dereference pointers to structs
In example3 we have a `*u` which dereferences a pointer to get to the
underlying struct value. This is good to know but I rarely ever do this.
Instead if a have a pointer variable `u` and I want the `name` property
I will use `u.name` which automatically deference the pointer.

I show this here because it's about this point in the training that
students are envisioning a life of frequent manual pointer dereference
but the reality is more convenient.

I also tweaked the formatting a bit so the output of example3 looks like

```
0xc42008e180    {name:Bill email:bill@ardanlabs.com logins:0}
Name: "Bill" Email: "bill@ardanlabs.com" Logins: 0

&logins[0xc4200a4020] logins[0xc42008e1a0] *logins[1]

0xc42008e180    {name:Bill email:bill@ardanlabs.com logins:1}
Name: "Bill" Email: "bill@ardanlabs.com" Logins: 1

```